Single-agent bugfixes, dungeon sweeps. - do not render metric stats when enjoying, by default (it's very slow) - add basic dungeon(2) sweeps - default to single-agent script train.py script in hyperparam sweeps if not otherwise specified - bugfix incorrect global timestep logging, and prevent integer overflow here

+++ purejaxrl/experimental/s5/wrappers.py
@@ -71,7 +71,7 @@ class LogEnvState:
     episode_lengths: int
     returned_episode_returns: float
     returned_episode_lengths: int
-    timestep: int
+    timestep: np.uint64
 
 class LogWrapper(GymnaxWrapper):
     """Log the episode returns and lengths."""
